Fetch fractional overlap for each neuropil for all images in BrainBase

Fetch fractional overlap for each neuropil for all images in BrainBase

neuropil_overlaps(nps, raw = FALSE, normalised_by = c("neuropil",
  "staining"))

Arguments

nps
The neuropils to fetch, specified either as integer ids or as the character vector abbreviated neuropil name (see neuropils in both cases. The special values of "brain" and "vnc" imply all brain or vnc neuropils, respectively.

raw
Whether to return the raw httr result

normalised_by
How to normalise the results. This denominator can either be the total number of voxels in each neuropil region or the total amount of staining in the whole expression pattern.

Value

processed data.frame with columns

imgId Integer image id

name Image name

... additional columns expressing the fraction of each neuropil domain occupied by foreground voxels and titled with the abbreviated neuropil names.

See also

The neuropils data.frame contains details of the different neuropil regions.

Examples

x=neuropil_overlaps(1152)
y=neuropil_overlaps(1152, normalised_by="staining")
# NOT RUN {
r=neuropil_overlaps(1152, raw=TRUE)
json_content=content(r)

all_brain=neuropil_overlaps('brain')
all_vnc=neuropil_overlaps('vnc')
# }